Update ED.drc.r
The proposed change would allow the user to enter either a function or a variance-covariance matrix. This is useful to be able e.g. to use the vcovCL() function, which requires a 'cluster' argument. At the moment, the 'cluster' argument is not accepted by the ED() function, but, once the proposed change is incorporated, the user will be able to use the vcovCL() function to calculate a cluster robust variance-covariance matrix and pass it to the ED() function as an argument.